Move Android to com.joshuapinter.RNContactsModule format
This commit:
1. Temporarily comments out the JS used for Android as it’s not being
used on iOS and causes the iOS to not run. Currently working toward
unification
2. Updates .gitgnore
