+++ android/build.gradle
@@ -16,15 +16,6 @@ android {
     lintOptions {
        warning 'InvalidPackage'
     }
-
-    buildTypes{
-      debug {
-
-      }
-      release {
-          
-      }
-    }
 }
 
 dependencies {
