--- android/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sat Mar 17 11:09:46 CET 2018
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-4.1-all.zip
