+++ android/src/main/AndroidManifest.xml
@@ -1,5 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.joshuapinter">
-
+          package="com.joshuapinter.RNUnifiedContacts">
 </manifest>
-
